ACS2020_V1

Calculate the American Cancer Society 2020 dietary index, using the daily serving sizes of foods consumed to calculate the vegetable, fruit, whole grain, red and processed meat, and sugar-sweetened beverages components. This version uses the percent of daily calories from highly processed foods and refined grains to calculate that component’s scores and is the preferred method of calculating the ACS 2020 diet score.

Usage

ACS2020_V1(
  SERV_DATA,
  RESPONDENTID,
  GENDER,
  VEG_SERV_ACS2020,
  VEG_ITEMS_SERV_ACS2020,
  FRT_SERV_ACS2020,
  FRT_ITEMS_SERV_ACS2020,
  WGRAIN_SERV_ACS2020,
  REDPROC_MEAT_SERV_ACS2020,
  HPFRG_RATIO_SERV_ACS2020,
  SSB_FRTJ_SERV_ACS2020
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

GENDER

The gender for each participant, 1=male, and 2=female

VEG_SERV_ACS2020

The serving size of All vegetable except potatoes and starchy vegetable, unit=servings/day (0.5 c of vege; 1 cup of green leafy (1 cup = 236.59 g)

VEG_ITEMS_SERV_ACS2020

The total number of unique vegetables (e.g., vegetable line items) reported by the participant. For example, a report of 3 servings of lettuce, 1 serving of kale, and 0.5 servings of broccoli would count as 3 vegetable line items.

FRT_SERV_ACS2020

The serving size of All whole fruits and no fruit juice, unit=servings/day (0.5 c of berries; 1 cup other fruits=236.59 g; 1 med fruit; 0.5 medium avocado)

FRT_ITEMS_SERV_ACS2020

The number of unique fruits (e.g., fruit line items) asked about on that survey/app (measuring the variety of fruits). For example, 3 serving of apple, 3 servings of banana, and 3 servings of blueberry are just 3 total number line items

WGRAIN_SERV_ACS2020

The serving size of whole grains, unit=grams/day

REDPROC_MEAT_SERV_ACS2020

The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=servings/day; 1 srv= 4 oz. unprocessed meat; 1.5 oz. processed meat (1 oz. = 28.35 g)

HPFRG_RATIO_SERV_ACS2020

The ratio of calories from highly processed foods and refined grains to the total daily calories (e.g. 35 % calories from HPF and ref grains), note: the ultra-processed variable for the score should not double count foods included in other parts of the score, for example, sugar-sweetened beverages or processed meats

SSB_FRTJ_SERV_ACS2020

The serving size of sugar-sweetened beverages and non-100% fruit juice, unit=servings/day = 1 ser= 8oz (1 oz. = 28.35 g)

Value

The ACS2020_V1 index/score and its component scores