AHEIP Calculation

Calculate the AHEIP dietary index (serving size-based), Mediterranean, using given the serving sizes of foods and nutrients consumed per 1 day

Usage

AHEIP(
  SERV_DATA,
  RESPONDENTID,
  VEG_SERV_AHEIP,
  FRT_SERV_AHEIP,
  WHITERED_RT_SERV_AHEIP,
  FIBER_SERV_AHEIP,
  TRANS_SERV_AHEIP,
  POLYSAT_RT_SERV_AHEIP,
  CALCIUM_SERV_AHEIP,
  FOLATE_SERV_AHEIP,
  IRON_SERV_AHEIP
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

VEG_SERV_AHEIP

The serving size of All vegetable except potatoes and legume, unit=servings/day (0.5 c of vege; 1 cup of green leafy (1 cup = 236.59 g)

FRT_SERV_AHEIP

The serving size of All whole fruits and no fruit juice, unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit (1 cup = 236.59 g)

WHITERED_RT_SERV_AHEIP

The serving size of the ratio of white and red meats, White meat = poultry + all fish, red meat = pork + beef + lamb + all organ meats + processed meat, unit=servings/day; 1 srv= 4 oz. unprocessed meat; 1.5 oz. processed meat (1 oz. = 28.35 g)

FIBER_SERV_AHEIP

The serving size of fibers, unit=grams/day

TRANS_SERV_AHEIP

The serving size of trans fat, unit='% of energy'

POLYSAT_RT_SERV_AHEIP

The serving size of polyunsaturated/saturated fats, unit=ratio

CALCIUM_SERV_AHEIP

The serving size of calcium, unit=mg/day

FOLATE_SERV_AHEIP

The serving size of folate, unit=mcg/day

IRON_SERV_AHEIP

The serving size of iron, unit=mg/day

Value

The AHEIP index/score

Examples

data("AHEIP_VALIDATION")
AHEIP(SERV_DATA = AHEIP_VALIDATION, RESPONDENTID = AHEIP_VALIDATION$id, VEG_SERV_AHEIP = AHEIP_VALIDATION$vegetable, FRT_SERV_AHEIP = AHEIP_VALIDATION$whole_fruit, WHITERED_RT_SERV_AHEIP = AHEIP_VALIDATION$white_meat_red_meat, FIBER_SERV_AHEIP = AHEIP_VALIDATION$fiber, TRANS_SERV_AHEIP = AHEIP_VALIDATION$trans_fat, POLYSAT_RT_SERV_AHEIP = AHEIP_VALIDATION$poly_fat_sat_fat, CALCIUM_SERV_AHEIP = AHEIP_VALIDATION$calcium, FOLATE_SERV_AHEIP = AHEIP_VALIDATION$folate, IRON_SERV_AHEIP = AHEIP_VALIDATION$iron)
#> # A tibble: 11 × 11
#>    RESPONDENTID AHEIP_ALL AHEIP_VEG AHEIP_FRT AHEIP_WHITEREAD AHEIP_FIBER
#>           <dbl>     <dbl>     <dbl>     <dbl>           <dbl>       <dbl>
#>  1            1        10         0         0               0           0
#>  2            2        17         1         1               1           1
#>  3            3        24         2         2               2           2
#>  4            4        31         3         3               3           3
#>  5            5        38         4         4               4           4
#>  6            6        45         5         5               5           5
#>  7            7        52         6         6               6           6
#>  8            8        59         7         7               7           7
#>  9            9        66         8         8               8           8
#> 10           10        73         9         9               9           9
#> 11           11        80        10        10              10          10
#> # ℹ 5 more variables: AHEIP_TRANS <dbl>, AHEIP_POLYSAT <dbl>,
#> #   AHEIP_CALCIUM <dbl>, AHEIP_FOLATE <dbl>, AHEIP_IRON <dbl>