AHEI_M_ASA24

Calculate the AHEI (male only) for the ASA24 data within 1 step

Usage

AHEI_M_ASA24(DATA_PATH, SSB_code = NULL, RECALL_SUMMARIZE = TRUE)

Arguments

DATA_PATH

The file path for the data. The file name should be like: Items.csv

SSB_code

The food code for sugar sweetened beverage, default is the SSB code from 17-18 FNDDS file

RECALL_SUMMARIZE

Whether to summarize the recalls to the participant level, default is TRUE

Value

The AHEI and its component scores

Examples

data("ASA24_exp_detailed")
AHEI_M_ASA24(ASA24_exp_detailed, SSB_code = NULL, RECALL_SUMMARIZE = TRUE) # for participants who are all male
#> Since no SSB code is provided, the default SSB code from 17-18 FNDDS file is used.
#> RECALL_SUMMARIZE = TRUE, summarizing HEI2015 for ASA24 data by averaging over all possible recalls per person per day...
#> Reminder: this AHEI index is for male only. Please stratify your data first and provide male only data.
#> # A tibble: 21 × 14
#> # Groups:   UserName, UserID [21]
#>    UserName UserID          AHEI_ALL AHEI_NOETOH AHEI_VEG AHEI_FRT AHEI_WGRAIN_M
#>    <chr>    <chr>              <dbl>       <dbl>    <dbl>    <dbl>         <dbl>
#>  1 THR01    e6541891-8a22-…     51.1        48.6    5.36     0.988         5.62 
#>  2 THR02    dbb278b0-b589-…     44.9        42.4    6.94     0             0.251
#>  3 THR03    f0115426-e6f8-…     39.0        36.5    9.24     2.48          0.857
#>  4 THR04    337d84a7-3bff-…     52.1        49.6    3.84     0             3.17 
#>  5 THR05    998f098a-8584-…     45.7        36.3    0        0             1.72 
#>  6 THR06    aae330db-8f22-…     45.1        42.6    3.16     2.56          0    
#>  7 THR07    76e0f3e1-8e17-…     51.2        48.7    7.66     1.58          2.70 
#>  8 THR08    24de994f-e06a-…     75.3        65.3    2.62     2.97         10    
#>  9 THR09    8f61ae64-6a82-…     48.2        45.7    0.202    3.16          1.75 
#> 10 THR10    b47ab1b8-e62b-…     58.6        56.1    5.27     0            10    
#> # ℹ 11 more rows
#> # ℹ 7 more variables: AHEI_NUTSLEG <dbl>, AHEI_N3FAT <dbl>, AHEI_PUFA <dbl>,
#> #   AHEI_SSB_FRTJ <dbl>, AHEI_REDPROC_MEAT <dbl>, AHEI_SODIUM <dbl>,
#> #   AHEI_ALCOHOL_M <dbl>