DASH_NHANES_FPED

Calculate the DASH for the NHANES_FPED data (after 2005) within 1 step

Usage

DASH_NHANES_FPED(
  FPED_IND_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  FPED_IND_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  SKIM_MILK_code = NULL,
  LOWF_CHEESE_code = NULL,
  SSB_code = NULL
)

Arguments

FPED_IND_PATH

The file path for the FPED IND data. The file name should be like: fped_dr1iff.sas7bdat

NUTRIENT_IND_PATH

The file path for the NUTRIENT IND data. The file name should be like: DR1IFF_J

FPED_IND_PATH2

The file path for the FPED IND data for day 2. The file name should be like: fped_dr2iff.sas7bdat

NUTRIENT_IND_PATH2

The file path for the NUTRIENT IND data for day 2. The file name should be like: DR2IFF_J

SKIM_MILK_code

The food code for skim milk. The default food codes are from 17-18 FNDDS file.

LOWF_CHEESE_code

The food code for low-fat cheese. The default food codes are from 17-18 FNDDS file.

SSB_code

The food code for sugar-sweetened beverages (SSB). The default food codes are from 17-18 FNDDS file.

Value

The DASH and its component scores and serving sizes

Examples

data("NHANES_20172018")
DASH_NHANES_FPED(NHANES_20172018$FPED_IND, NHANES_20172018$NUTRIENT_IND, NHANES_20172018$FPED_IND2, NHANES_20172018$NUTRIENT_IND2)
#> Since no SSB code is provided, the default SSB code from 17-18 FNDDS file is used.
#> Since no skim milk code is provided, the default skim milk code from 17-18 FNDDS file is used.
#> Since no low-fat cheese code is provided, the default low-fat cheese code from 17-18 FNDDS file is used.
#> Reminder: this DASH index uses quintiles to rank participants' food/drink serving sizes and then calculate DASH component scores, which may generate results that are specific to your study population but not comparable to other populations.
#> # A tibble: 6,488 × 10
#>     SEQN DASH_ALL DASH_FRT DASH_VEG DASH_NUTSLEG DASH_WGRAIN DASH_LOWF_DAIRY
#>    <dbl>    <dbl>    <dbl>    <dbl>        <dbl>       <dbl>           <dbl>
#>  1 93704     23.5      4.5      1            1           2.5               1
#>  2 93705     18.5      1.5      3            4           1                 1
#>  3 93707     20        1        2            2           4.5               1
#>  4 93708     23        2        3.5          3           2                 1
#>  5 93710     25.5      3.5      1.5          2.5         3.5               1
#>  6 93711     31        4.5      4.5          5           3                 5
#>  7 93712     18        3        3.5          3           3                 1
#>  8 93713     24.5      4.5      4.5          2.5         2.5               1
#>  9 93714     17.5      2.5      1            1           3                 3
#> 10 93715     14.5      2        1.5          1           1                 1
#> # ℹ 6,478 more rows
#> # ℹ 3 more variables: DASH_SODIUM <dbl>, DASH_REDPROC_MEAT <dbl>,
#> #   DASH_SSB_FRTJ <dbl>