DASH_NHANES_MPED

Calculate the DASH for the NHANES_MPED data (before 2005, 1999-2004) within 1 step for day 1, day 2, or day 1 and 2 combined (age >= 2 only)

Usage

DASH_NHANES_MPED(
  MPED_PER_100_GRAM_PATH = NULL,
  WJFRT = NULL,
  NUTRIENT_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  DEMO_PATH,
  NUTRIENT_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  SKIM_MILK_code = NULL,
  LOWF_CHEESE_code = NULL,
  SSB_code = NULL
)

Arguments

MPED_PER_100_GRAM_PATH

The file path for the MPED per 100 gram data for the day 1 and day 2 data. The file name should be like: pyr_tot_d1.sas7bdat

WJFRT

The file path for the WJFRT data for the day 1 and day2 data. The file name should be like: wjfrt.sas7bdat

NUTRIENT_PATH

The file path for the NUTRIENT data for the day 1 data. The file name should be like: DR1TOT_J.XPT or DRXTOT_B.XPT

NUTRIENT_IND_PATH

The file path for the NUTRIENT_IND data for the day 1 data. The file name should be like: DR1IFF_J.XPT

DEMO_PATH

The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT

NUTRIENT_PATH2

The file path for the NUTRIENT2 data for the day 2 data. The file name should be like: DR2TOT_J.XPT

NUTRIENT_IND_PATH2

The file path for the NUTRIENT_IND2 data for the day 2 data The file name should be like: DR2IFF_J.XPT

SKIM_MILK_code

The food code for skim milk. The default food codes are from 17-18 FNDDS file.

LOWF_CHEESE_code

The food code for low-fat cheese. The default food codes are from 17-18 FNDDS file.

SSB_code

The food code for sugar-sweetened beverages (SSB). The default food codes are from 17-18 FNDDS file.

Value

The DASH and its component scores and serving sizes

Examples

data("NHANES_20032004")
DASH_NHANES_MPED(MPED_PER_100_GRAM_PATH = NHANES_20032004$MPED_PER_100_GRAM, WJFRT = NHANES_20032004$WJFRT, NUTRIENT_PATH = NHANES_20032004$NUTRIENT, NUTRIENT_IND_PATH = NHANES_20032004$NUTRIENT_IND, DEMO_PATH = NHANES_20032004$DEMO, NUTRIENT_PATH2 = NHANES_20032004$NUTRIENT2, NUTRIENT_IND_PATH2 = NHANES_20032004$NUTRIENT_IND2)
#> Since no SSB code is provided, the default SSB code from 17-18 FNDDS file is used.
#> Since no skim milk code is provided, the default skim milk code from 17-18 FNDDS file is used.
#> Since no low-fat cheese code is provided, the default low-fat cheese code from 17-18 FNDDS file is used.
#> Reminder: this DASH index uses quintiles to rank participants' food/drink serving sizes and then calculate DASH component scores, which may generate results that are specific to your study population but not comparable to other populations.
#> # A tibble: 7,650 × 10
#>     SEQN DASH_ALL DASH_FRT DASH_VEG DASH_NUTSLEG DASH_WGRAIN DASH_LOWF_DAIRY
#>    <dbl>    <dbl>    <dbl>    <dbl>        <dbl>       <dbl>           <dbl>
#>  1 21005     15.5      2        2            2.5         3                 1
#>  2 21006     18        2        2            2.5         2.5               1
#>  3 21007     18        3        3            1           1                 1
#>  4 21008     17.5      1        3.5          2.5         3                 1
#>  5 21009     19.5      2        3            2           2                 1
#>  6 21010     18        1.5      2            4           1                 1
#>  7 21012     17.5      1        3            4           2.5               1
#>  8 21013     16        2.5      2            2           1                 1
#>  9 21014     25.5      3        4            4           2.5               1
#> 10 21015     27.5      3        3            4           4.5               1
#> # ℹ 7,640 more rows
#> # ℹ 3 more variables: DASH_SODIUM <dbl>, DASH_REDPROC_MEAT <dbl>,
#> #   DASH_SSB_FRTJ <dbl>