DII_NHANES_MPED

Calculate the DII for the NHANES_MPED data (before 2005, 1999-2004) within 1 step for day 1, day 2, or day 1 and 2 combined (age >= 2 only)

Usage

DII_NHANES_MPED(
  MPED_PER_100_GRAM_PATH = NULL,
  WJFRT = NULL,
  NUTRIENT_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  DEMO_PATH,
  NUTRIENT_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  OTHER_INGREDIENTS1 = NULL,
  OTHER_INGREDIENTS2 = NULL
)

Arguments

MPED_PER_100_GRAM_PATH

The file path for the MPED per 100 gram data for the day 1 and day 2 data. The file name should be like: pyr_tot_d1.sas7bdat

WJFRT

The file path for the WJFRT data for the day 1 and day2 data. The file name should be like: wjfrt.sas7bdat

NUTRIENT_PATH

The file path for the NUTRIENT data for the day 1 data. The file name should be like: DR1TOT_J.XPT or DRXTOT_B.XPT

NUTRIENT_IND_PATH

The file path for the NUTRIENT_IND data for the day 1 data. The file name should be like: DR1IFF_J.XPT

DEMO_PATH

The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT

NUTRIENT_PATH2

The file path for the NUTRIENT2 data for the day 2 data. The file name should be like: DR2TOT_J.XPT

NUTRIENT_IND_PATH2

The file path for the NUTRIENT_IND2 data for the day 2 data The file name should be like: DR2IFF_J.XPT

OTHER_INGREDIENTS1

The file path for the DII other ingredients data for the day 1 data.

OTHER_INGREDIENTS2

The file path for the DII other ingredients data for the day 2 data.

Value

The DII and its component scores and serving sizes

Examples

data("NHANES_20032004")
data("DII_OTHER_INGREDIENTS_day1")
data("DII_OTHER_INGREDIENTS_day2")
DII_NHANES_MPED(MPED_PER_100_GRAM_PATH = NHANES_20032004$MPED_PER_100_GRAM, WJFRT = NHANES_20032004$WJFRT, NUTRIENT_PATH = NHANES_20032004$NUTRIENT, NUTRIENT_IND_PATH = NHANES_20032004$NUTRIENT_IND, DEMO_PATH = NHANES_20032004$DEMO, NUTRIENT_PATH2 = NHANES_20032004$NUTRIENT2, NUTRIENT_IND_PATH2 = NHANES_20032004$NUTRIENT_IND2, OTHER_INGREDIENTS1 = DII_OTHER_INGREDIENTS_day1, OTHER_INGREDIENTS2 = DII_OTHER_INGREDIENTS_day2)
#> VITD is not included in the calculation in the first day of NHANES data.
#> It is normal to see warnings if you do not provide all arguments using DII. The algorithm will only count the arguments you enter to calculate the DII. All warnings are about the first column you don't use. For example, if you only entered alcohol, vitamin b12, and vitamin b6, all warnings would remind you that bcarotene is not found.
#> VITD is not included in the calculation in the first day of NHANES data.
#> It is normal to see warnings if you do not provide all arguments using DII. The algorithm will only count the arguments you enter to calculate the DII. All warnings are about the first column you don't use. For example, if you only entered alcohol, vitamin b12, and vitamin b6, all warnings would remind you that bcarotene is not found.
#> Reminder: This function does not use all the original DII variables. Eugenol, trans fat, turmeric, Green/black tea, Rosemary are not included.
#> Day 1 and Day 2 data are used for the calculation.
#> Note: Flavonoid data is only available for NHANES 2007-2010 and 2017-2018.
#> # A tibble: 7,647 × 48
#>     SEQN DII_ALL DII_NOETOH ALCOHOL_DII VITB12_DII VITB6_DII BCAROTENE_DII
#>    <dbl>   <dbl>      <dbl>       <dbl>      <dbl>     <dbl>         <dbl>
#>  1 21005    2.68      2.40        0.278    -0.0305   -0.107          0.555
#>  2 21006    4.76      4.48        0.278    -0.0868    0.224          0.563
#>  3 21007    4.82      4.54        0.278    -0.0904    0.277          0.540
#>  4 21008    4.28      4.00        0.278    -0.0383    0.150          0.537
#>  5 21009    1.21      0.937       0.278     0.0337   -0.293          0.528
#>  6 21010    3.34      3.61       -0.274    -0.0499    0.0303         0.563
#>  7 21012    3.23      3.51       -0.278     0.0360   -0.194          0.549
#>  8 21013    4.57      4.29        0.278    -0.0565    0.150          0.559
#>  9 21014    1.93      1.65        0.278     0.0706   -0.204          0.524
#> 10 21015    2.39      2.11        0.278    -0.0551   -0.0627         0.134
#> # ℹ 7,637 more rows
#> # ℹ 41 more variables: CAFFEINE_DII <dbl>, CARB_DII <dbl>, CHOLES_DII <dbl>,
#> #   KCAL_DII <dbl>, EUGENOL_DII <dbl>, TOTALFAT_DII <dbl>, FIBER_DII <dbl>,
#> #   FOLICACID_DII <dbl>, GARLIC_DII <dbl>, GINGER_DII <dbl>, IRON_DII <dbl>,
#> #   MG_DII <dbl>, MUFA_DII <dbl>, NIACIN_DII <dbl>, N3FAT_DII <dbl>,
#> #   N6FAT_DII <dbl>, ONION_DII <dbl>, PROTEIN_DII <dbl>, PUFA_DII <dbl>,
#> #   RIBOFLAVIN_DII <dbl>, SAFFRON_DII <dbl>, SATFAT_DII <dbl>, SE_DII <dbl>, …