DI_GM Calculation

Calculate the Dietary Index for Gut Microbiota (DI-GM), using given the serving sizes of foods and nutrients consumed per 1 day

Usage

DI_GM(
  SERV_DATA,
  RESPONDENTID,
  GENDER,
  AVOCADO,
  BROCCOLI,
  CHICKPEA,
  COFFEE,
  CRANBERRY,
  FERMENTED_DAIRY,
  FIBER,
  GREEN_TEA,
  SOYBEAN,
  WHOLE_GRAIN,
  TOTAL_FAT_PERCENTAGE,
  PROCESSED_MEAT,
  RED_MEAT,
  REFINED_GRAIN
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

GENDER

The gender of the participant

AVOCADO

The serving size of Avocado. Unit does not matter here since the scoring is based on the median intake of the food in the population. However, it is important to keep the same unit for all indivudal food consumption data.

BROCCOLI

The serving size of Broccoli. Unit does not matter here as indicated previously.

CHICKPEA

The serving size of Chickpea. Unit does not matter here as indicated previously.

COFFEE

The serving size of Coffee. Unit does not matter here as indicated previously.

CRANBERRY

The serving size of Cranberry. Unit does not matter here as indicated previously.

FERMENTED_DAIRY

The serving size of Fermented Dairy. Unit does not matter here as indicated previously.

FIBER

The serving size of Fiber. Unit does not matter here as indicated previously.

GREEN_TEA

The serving size of Green Tea. Unit does not matter here as indicated previously.

SOYBEAN

The serving size of Soybean. Unit does not matter here as indicated previously.

WHOLE_GRAIN

The serving size of Whole Grain. Unit does not matter here as indicated previously.

TOTAL_FAT_PERCENTAGE

The total fat calorie percentage of the total calories consumed. Unit is percent calories.

PROCESSED_MEAT

The serving size of Processed Meat. Unit does not matter here as indicated previously.

RED_MEAT

The serving size of Red Meat. Unit does not matter here as indicated previously.

REFINED_GRAIN

The serving size of Refined Grain. Unit does not matter here as indicated previously.

Value

The DI_GM index/score

Examples

data("DI_GM_VALIDATION")
DI_GM(SERV_DATA = DI_GM_VALIDATION, RESPONDENTID = DI_GM_VALIDATION$id, GENDER = DI_GM_VALIDATION$gender, AVOCADO = DI_GM_VALIDATION$avocado, BROCCOLI = DI_GM_VALIDATION$broccoli, CHICKPEA = DI_GM_VALIDATION$chickpea, COFFEE = DI_GM_VALIDATION$coffee, CRANBERRY = DI_GM_VALIDATION$cranberry, FERMENTED_DAIRY = DI_GM_VALIDATION$fermented_dairy, FIBER = DI_GM_VALIDATION$fiber, GREEN_TEA = DI_GM_VALIDATION$green_tea, SOYBEAN = DI_GM_VALIDATION$soybean, WHOLE_GRAIN = DI_GM_VALIDATION$whole_grain, TOTAL_FAT_PERCENTAGE = DI_GM_VALIDATION$total_fat_percentage, PROCESSED_MEAT = DI_GM_VALIDATION$processed_meat, RED_MEAT = DI_GM_VALIDATION$red_meat, REFINED_GRAIN = DI_GM_VALIDATION$refined_grain)
#> Reminder: this DI_GM index uses sex-specific medians to rank participants' food/drink serving sizes and then calculate DI_GM component scores, which may generate results that are specific to your study population but not comparable to other populations.
#> # A tibble: 10 × 17
#>    RESPONDENTID GENDER DI_GM_TOTAL DI_GM_AVOCADO DI_GM_BROCCOLI DI_GM_CHICKPEA
#>           <dbl>  <dbl>       <dbl>         <dbl>          <dbl>          <dbl>
#>  1            1      2           0             0              0              0
#>  2            2      2           0             0              0              0
#>  3            3      2          11             1              1              1
#>  4            4      2          14             1              1              1
#>  5            5      2          14             1              1              1
#>  6            6      1          14             1              1              1
#>  7            7      1          14             1              1              1
#>  8            8      1          11             1              1              1
#>  9            9      1           0             0              0              0
#> 10           10      1           0             0              0              0
#> # ℹ 11 more variables: DI_GM_COFFEE <dbl>, DI_GM_CRANBERRY <dbl>,
#> #   DI_GM_FERMENTED_DAIRY <dbl>, DI_GM_FIBER <dbl>, DI_GM_GREEN_TEA <dbl>,
#> #   DI_GM_SOYBEAN <dbl>, DI_GM_WHOLE_GRAIN <dbl>,
#> #   DI_GM_TOTAL_FAT_PERCENTAGE <dbl>, DI_GM_PROCESSED_MEAT <dbl>,
#> #   DI_GM_RED_MEAT <dbl>, DI_GM_REFINED_GRAIN <dbl>