DI_GM_NHANES_MPED

Calculate the Dietary Index for Gut Microbiota using a binary cutoff for the NHANES_MPED data (before 2005, 1999-2004) within 1 step for day 1, day 2, or day 1 and 2 combined (age >= 2 only)

Usage

DI_GM_NHANES_MPED(
  MPED_PER_100_GRAM_PATH = NULL,
  WJFRT = NULL,
  DEMO_PATH,
  NUTRIENT_PATH = NULL,
  NUTRIENT_IND_PATH = NULL,
  NUTRIENT_PATH2 = NULL,
  NUTRIENT_IND_PATH2 = NULL,
  AVOCADO_CODE = NULL,
  BROCCOLI_CODE = NULL,
  CHICKPEA_CODE = NULL,
  COFFEE_CODE = NULL,
  CRANBERRY_CODE = NULL,
  FERMENTED_DAIRY_CODE = NULL,
  GREEN_TEA_CODE = NULL,
  SOYBEAN_CODE = NULL
)

Arguments

MPED_PER_100_GRAM_PATH

The file path for the MPED per 100 gram data for the day 1 and day 2 data. The file name should be like: pyr_tot_d1.sas7bdat

WJFRT

The file path for the WJFRT data for the day 1 and day2 data. The file name should be like: wjfrt.sas7bdat

DEMO_PATH

The file path for the DEMOGRAPHIC data. The file name should be like: DEMO_J.XPT

NUTRIENT_PATH

The file path for the NUTRIENT data for the day 1 data. The file name should be like: DR1TOT_J.XPT or DRXTOT_B.XPT

NUTRIENT_IND_PATH

The file path for the NUTRIENT_IND data for the day 1 data. The file name should be like: DR1IFF_J.XPT

NUTRIENT_PATH2

The file path for the NUTRIENT2 data for the day 2 data. The file name should be like: DR2TOT_J.XPT

NUTRIENT_IND_PATH2

The file path for the NUTRIENT_IND2 data for the day 2 data The file name should be like: DR2IFF_J.XPT

AVOCADO_CODE

The code for avocado in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

BROCCOLI_CODE

The code for broccoli in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

CHICKPEA_CODE

The code for chickpea in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

COFFEE_CODE

The code for coffee in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

CRANBERRY_CODE

The code for cranberry in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

GREEN_TEA_CODE

The code for green tea in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

SOYBEAN_CODE

The code for soybean foods in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

FERMENT_DAIRY_CODE

The code for fermented dairy in the FPED data. The default food codes are from the FPED data for NHANES 2017-2018

Value

The DI_GM and its component scores and serving sizes

Examples

data("NHANES_20032004")
DI_GM_NHANES_MPED(MPED_PER_100_GRAM_PATH = NHANES_20032004$MPED_PER_100_GRAM, WJFRT = NHANES_20032004$WJFRT, NUTRIENT_PATH = NHANES_20032004$NUTRIENT, NUTRIENT_IND_PATH = NHANES_20032004$NUTRIENT_IND, DEMO_PATH = NHANES_20032004$DEMO, NUTRIENT_PATH2 = NHANES_20032004$NUTRIENT2, NUTRIENT_IND_PATH2 = NHANES_20032004$NUTRIENT_IND2)
#> The default food codes for avocado from 17-18 FNDDS file is used.
#> The default food codes for broccoli from 17-18 FNDDS file is used.
#> The default food codes for chickpea from 17-18 FNDDS file is used.
#> The default food codes for coffee from 17-18 FNDDS file is used.
#> The default food codes for cranberry from 17-18 FNDDS file is used.
#> The default food codes for fermented dairy from 17-18 FNDDS file is used.
#> The default food codes for green tea from 17-18 FNDDS file is used.
#> The default food codes for soybean from 17-18 FNDDS file is used.
#> Calculating the DI_GM total and component scores for the first day data...
#> Reminder: this DI_GM index uses sex-specific medians to rank participants' food/drink serving sizes and then calculate DI_GM component scores, which may generate results that are specific to your study population but not comparable to other populations.
#> Calculating the DI_GM total and component scores for the second day data...
#> Reminder: this DI_GM index uses sex-specific medians to rank participants' food/drink serving sizes and then calculate DI_GM component scores, which may generate results that are specific to your study population but not comparable to other populations.
#> # A tibble: 7,647 × 17
#>     SEQN RIAGENDR DI_GM_TOTAL DI_GM_AVOCADO DI_GM_BROCCOLI DI_GM_CHICKPEA
#>    <dbl>    <dbl>       <dbl>         <dbl>          <dbl>          <dbl>
#>  1 21005        1         3               0              0              0
#>  2 21006        2         3               0              0              0
#>  3 21007        2         2               0              0              0
#>  4 21008        1         2               0              0              0
#>  5 21009        1         4.5             0              0              0
#>  6 21010        2         4.5             0              0              0
#>  7 21012        1         3               0              0              0
#>  8 21013        2         2.5             0              0              0
#>  9 21014        1         4.5             0              0              0
#> 10 21015        1         6.5             0              0              0
#> # ℹ 7,637 more rows
#> # ℹ 11 more variables: DI_GM_COFFEE <dbl>, DI_GM_CRANBERRY <dbl>,
#> #   DI_GM_FERMENTED_DAIRY <dbl>, DI_GM_FIBER <dbl>, DI_GM_GREEN_TEA <dbl>,
#> #   DI_GM_SOYBEAN <dbl>, DI_GM_WHOLE_GRAIN <dbl>,
#> #   DI_GM_TOTAL_FAT_PERCENTAGE <dbl>, DI_GM_PROCESSED_MEAT <dbl>,
#> #   DI_GM_RED_MEAT <dbl>, DI_GM_REFINED_GRAIN <dbl>