HEI2015 Calculation

Calculate the HEI2015 dietary index, Healthy eating index 2015, using given the serving sizes of foods and nutrients consumed per 1 day

Usage

HEI2015(
  SERV_DATA,
  RESPONDENTID,
  TOTALKCAL_HEI2015,
  TOTALFRT_SERV_HEI2015,
  FRT_SERV_HEI2015,
  VEG_SERV_HEI2015,
  GREENNBEAN_SERV_HEI2015,
  TOTALPRO_SERV_HEI2015,
  SEAPLANTPRO_SERV_HEI2015,
  WHOLEGRAIN_SERV_HEI2015,
  DAIRY_SERV_HEI2015,
  FATTYACID_SERV_HEI2015,
  REFINEDGRAIN_SERV_HEI2015,
  SODIUM_SERV_HEI2015,
  ADDEDSUGAR_SERV_HEI2015,
  SATFAT_SERV_HEI2015
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

TOTALKCAL_HEI2015

The total calorie from all foods and drinks

TOTALFRT_SERV_HEI2015

The serving size of total fruits including fruit juice, unit= cup eq.

FRT_SERV_HEI2015

The serving size of Citrus, Melons, Berries + Other Intact Fruits, unit= cup eq.

VEG_SERV_HEI2015

The serving size of vegetables Total Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.

GREENNBEAN_SERV_HEI2015

The serving size of Dark Green Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.

TOTALPRO_SERV_HEI2015

The serving size of Total Meat, Poultry, and Seafood (including organ meats and cured meats) + Eggs + Nuts and Seeds + Soy + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz

SEAPLANTPRO_SERV_HEI2015

The serving size of Seafood (high in n-3) + Seafood (low in n-3) + Soy + Nuts and Seeds + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz

WHOLEGRAIN_SERV_HEI2015

The serving size of whole grains, unit=oz. eq.

DAIRY_SERV_HEI2015

The serving size of all dairy, unit=cup eq.

FATTYACID_SERV_HEI2015

The serving size of (Total Monounsaturated Fatty Acids + Total Polyunsaturated Fatty Acids)/Total Saturated Fatty Acids, unit=g

REFINEDGRAIN_SERV_HEI2015

The serving size of refined grains, unit=oz. eq.

SODIUM_SERV_HEI2015

The serving size of sodium, unit=mg

ADDEDSUGAR_SERV_HEI2015

The serving size of added sugar, unit=tsp

SATFAT_SERV_HEI2015

The serving size of Total Saturated Fatty Acids, unit=g

Value

The HEI2015 index/score

Examples

data("HEI2015_VALIDATION")
HEI2015(SERV_DATA = HEI2015_VALIDATION, RESPONDENTID = HEI2015_VALIDATION$id, TOTALKCAL_HEI2015 = HEI2015_VALIDATION$kcal, TOTALFRT_SERV_HEI2015 = HEI2015_VALIDATION$total_fruit, FRT_SERV_HEI2015 = HEI2015_VALIDATION$whole_fruit, VEG_SERV_HEI2015 = HEI2015_VALIDATION$total_vegetable, GREENNBEAN_SERV_HEI2015 = HEI2015_VALIDATION$green_and_bean, TOTALPRO_SERV_HEI2015 = HEI2015_VALIDATION$total_protein, SEAPLANTPRO_SERV_HEI2015 = HEI2015_VALIDATION$seafood_plant_protein, WHOLEGRAIN_SERV_HEI2015 = HEI2015_VALIDATION$whole_grain, DAIRY_SERV_HEI2015 = HEI2015_VALIDATION$dairy, FATTYACID_SERV_HEI2015 = HEI2015_VALIDATION$fatty_acid, REFINEDGRAIN_SERV_HEI2015 = HEI2015_VALIDATION$refined_grain, SODIUM_SERV_HEI2015 = HEI2015_VALIDATION$sodium, ADDEDSUGAR_SERV_HEI2015 = HEI2015_VALIDATION$added_sugar, SATFAT_SERV_HEI2015 = HEI2015_VALIDATION$saturated_fat)
#> # A tibble: 11 × 16
#>    RESPONDENTID TOTALKCAL_HEI2015 HEI2015_ALL HEI2015_TOTALFRT HEI2015_FRT
#>           <dbl>             <dbl>       <dbl>            <dbl>       <dbl>
#>  1            1              1000          40              0           0  
#>  2            2              1000          42              0.5         0.5
#>  3            3              1000          44              1           1  
#>  4            4              1000          46              1.5         1.5
#>  5            5              1000          48              2           2  
#>  6            6              1000          50              2.5         2.5
#>  7            7              1000          52              3           3  
#>  8            8              1000          54              3.5         3.5
#>  9            9              1000          56              4           4  
#> 10           10              1000          58              4.5         4.5
#> 11           11              1000          60              5           5  
#> # ℹ 11 more variables: HEI2015_VEG <dbl>, HEI2015_GREENNBEAN <dbl>,
#> #   HEI2015_TOTALPRO <dbl>, HEI2015_SEAPLANTPRO <dbl>,
#> #   HEI2015_WHOLEGRAIN <dbl>, HEI2015_DAIRY <dbl>, HEI2015_FATTYACID <dbl>,
#> #   HEI2015_REFINEDGRAIN <dbl>, HEI2015_SODIUM <dbl>, HEI2015_ADDEDSUGAR <dbl>,
#> #   HEI2015_SATFAT <dbl>