HEI2020_V2 Calculation

Calculate the HEI2020 dietary index, Healthy eating index 2020 and HEI-Toddlers-2020, using given the serving sizes of foods and nutrients consumed per 1 day. This version has the Added sugar serving size as tsp (teaspoon) and the saturated fat serving size as g (gram), instead of the

Usage

HEI2020_V2(
  SERV_DATA,
  RESPONDENTID,
  AGE,
  TOTALKCAL_HEI2020,
  TOTALFRT_SERV_HEI2020,
  FRT_SERV_HEI2020,
  VEG_SERV_HEI2020,
  GREENNBEAN_SERV_HEI2020,
  TOTALPRO_SERV_HEI2020,
  SEAPLANTPRO_SERV_HEI2020,
  WHOLEGRAIN_SERV_HEI2020,
  DAIRY_SERV_HEI2020,
  FATTYACID_SERV_HEI2020,
  REFINEDGRAIN_SERV_HEI2020,
  SODIUM_SERV_HEI2020,
  ADDEDSUGAR_SERV_HEI2020,
  SATFAT_SERV_HEI2020
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

AGE

The age of the participant in years

TOTALKCAL_HEI2020

The total calorie from all foods and drinks

TOTALFRT_SERV_HEI2020

The serving size of total fruits including fruit juice, unit= cup eq.

FRT_SERV_HEI2020

The serving size of Citrus, Melons, Berries + Other Intact Fruits, unit= cup eq.

VEG_SERV_HEI2020

The serving size of vegetables Total Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.

GREENNBEAN_SERV_HEI2020

The serving size of Dark Green Vegetables + Legumes (Beans and Peas) in cup equivalents, unit= cup eq.

TOTALPRO_SERV_HEI2020

The serving size of Total Meat, Poultry, and Seafood (including organ meats and cured meats) + Eggs + Nuts and Seeds + Soy + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz

SEAPLANTPRO_SERV_HEI2020

The serving size of Seafood (high in n-3) + Seafood (low in n-3) + Soy + Nuts and Seeds + Legumes (Beans and Peas) in oz equivalents, unit=oz. eq., 1 cup legume = 4 oz

WHOLEGRAIN_SERV_HEI2020

The serving size of whole grains, unit=oz. eq.

DAIRY_SERV_HEI2020

The serving size of all dairy, unit=cup eq.

FATTYACID_SERV_HEI2020

The serving size of (Total Monounsaturated Fatty Acids + Total Polyunsaturated Fatty Acids)/Total Saturated Fatty Acids, unit=g

REFINEDGRAIN_SERV_HEI2020

The serving size of refined grains, unit=oz. eq.

SODIUM_SERV_HEI2020

The serving size of sodium, unit=mg

ADDEDSUGAR_SERV_HEI2020

The serving size of added sugar, unit=tsp

SATFAT_SERV_HEI2020

The serving size of Total Saturated Fatty Acids, unit=g

Value

The HEI2020 index/score

Examples

data("HEI2020_V2_VALIDATION")
HEI2020_V2(SERV_DATA = HEI2020_V2_VALIDATION, RESPONDENTID = HEI2020_V2_VALIDATION$id, AGE = HEI2020_V2_VALIDATION$age, TOTALKCAL_HEI2020 = HEI2020_V2_VALIDATION$kcal, TOTALFRT_SERV_HEI2020 = HEI2020_V2_VALIDATION$total_fruit, FRT_SERV_HEI2020 = HEI2020_V2_VALIDATION$whole_fruit, VEG_SERV_HEI2020 = HEI2020_V2_VALIDATION$total_vegetable, GREENNBEAN_SERV_HEI2020 = HEI2020_V2_VALIDATION$green_and_bean, TOTALPRO_SERV_HEI2020 = HEI2020_V2_VALIDATION$total_protein, SEAPLANTPRO_SERV_HEI2020 = HEI2020_V2_VALIDATION$seafood_plant_protein, WHOLEGRAIN_SERV_HEI2020 = HEI2020_V2_VALIDATION$whole_grain, DAIRY_SERV_HEI2020 = HEI2020_V2_VALIDATION$dairy, FATTYACID_SERV_HEI2020 = HEI2020_V2_VALIDATION$fatty_acid, REFINEDGRAIN_SERV_HEI2020 = HEI2020_V2_VALIDATION$refined_grain, SODIUM_SERV_HEI2020 = HEI2020_V2_VALIDATION$sodium, ADDEDSUGAR_SERV_HEI2020 = HEI2020_V2_VALIDATION$added_sugar, SATFAT_SERV_HEI2020 = HEI2020_V2_VALIDATION$saturated_fat)
#> # A tibble: 22 × 17
#>    RESPONDENTID   AGE TOTALKCAL_HEI2020 HEI2020_ALL HEI2020_TOTALFRT HEI2020_FRT
#>           <dbl> <dbl>             <dbl>       <dbl>            <dbl>       <dbl>
#>  1            1    18              1000        40                0           0  
#>  2            2    18              1000        42                0.5         0.5
#>  3            3    18              1000        44.0              1           1  
#>  4            4    18              1000        46.0              1.5         1.5
#>  5            5    18              1000        48.0              2           2  
#>  6            6    18              1000        50.0              2.5         2.5
#>  7            7    18              1000        52                3           3  
#>  8            8    18              1000        54                3.5         3.5
#>  9            9    18              1000        56                4           4  
#> 10           10    18              1000        58                4.5         4.5
#> # ℹ 12 more rows
#> # ℹ 11 more variables: HEI2020_VEG <dbl>, HEI2020_GREENNBEAN <dbl>,
#> #   HEI2020_TOTALPRO <dbl>, HEI2020_SEAPLANTPRO <dbl>,
#> #   HEI2020_WHOLEGRAIN <dbl>, HEI2020_DAIRY <dbl>, HEI2020_FATTYACID <dbl>,
#> #   HEI2020_REFINEDGRAIN <dbl>, HEI2020_SODIUM <dbl>, HEI2020_ADDEDSUGAR <dbl>,
#> #   HEI2020_SATFAT <dbl>