MED Calculation

Calculate the MED dietary index, Mediterranean, using given the serving sizes of foods and nutrients consumed per 1 day

Usage

MED(
  SERV_DATA,
  RESPONDENTID,
  FRT_FRTJ_SERV_MED,
  VEG_SERV_MED,
  WGRAIN_SERV_MED,
  LEGUMES_SERV_MED,
  NUTS_SERV_MED,
  FISH_SERV_MED,
  REDPROC_MEAT_SERV_MED,
  MONSATFAT_SERV_MED,
  ALCOHOL_SERV_MED
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

FRT_FRTJ_SERV_MED

The serving size of All fruits and 100% fruit juices, unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit (1 cup = 236.59 g); 1 cup fruit juice

VEG_SERV_MED

The serving size of All vegetables except potatoes and legumes, unit=0.5 c of vege; 1 cup of green leafy

WGRAIN_SERV_MED

The serving size of whole grains, including Whole-grain ready-to-eat cereals, cooked cereals, crackers, dark breads, brown rice, other grains, wheat germ, bran, popcorn, unit=1oz

LEGUMES_SERV_MED

The serving size of legumes, including Tofu, string beans, peas, beans, unit=*oz, 1 cup legume = 4 oz

NUTS_SERV_MED

The serving size of nuts, including Nuts, peanut butter, unit=1oz

FISH_SERV_MED

The serving size of all fish, including Fish and shrimp, breaded fish, unit=4oz

REDPROC_MEAT_SERV_MED

The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=serving (4 oz. unprocessed meat; 1.5 oz. processed meat)

MONSATFAT_SERV_MED

The serving size of the ratio of monounsaturated fat to saturated fat, unit=ratio

ALCOHOL_SERV_MED

The serving size of alcohol, unit=13g

Value

The MED index/score

Examples

data("MED_VALIDATION")
MED(MED_VALIDATION, RESPONDENTID = MED_VALIDATION$id, FRT_FRTJ_SERV_MED = MED_VALIDATION$fruit, VEG_SERV_MED = MED_VALIDATION$vegetable, WGRAIN_SERV_MED = MED_VALIDATION$whole_grain, LEGUMES_SERV_MED = MED_VALIDATION$legume, NUTS_SERV_MED = MED_VALIDATION$nut, FISH_SERV_MED = MED_VALIDATION$fish, REDPROC_MEAT_SERV_MED = MED_VALIDATION$red_processed_meat, MONSATFAT_SERV_MED = MED_VALIDATION$monofat_satfat, ALCOHOL_SERV_MED = MED_VALIDATION$alcohol)
#> Reminder: this MED index uses medians to rank participants' food/drink serving sizes and then calculate MED component scores, which may generate results that are specific to your study population but not comparable to other populations.
#> # A tibble: 11 × 12
#>    RESPONDENTID MED_ALL MED_NOETOH MED_FRT MED_VEG MED_WGRAIN MED_LEGUMES
#>           <dbl>   <dbl>      <dbl>   <dbl>   <dbl>      <dbl>       <dbl>
#>  1            1       1          1       0       0          0           0
#>  2            2       1          1       0       0          0           0
#>  3            3       1          1       0       0          0           0
#>  4            4       1          1       0       0          0           0
#>  5            5       2          1       0       0          0           0
#>  6            6       8          7       1       1          1           1
#>  7            7       8          7       1       1          1           1
#>  8            8       8          7       1       1          1           1
#>  9            9       8          7       1       1          1           1
#> 10           10       7          7       1       1          1           1
#> 11           11       7          7       1       1          1           1
#> # ℹ 5 more variables: MED_NUTS <dbl>, MED_FISH <dbl>, MED_REDPROC_MEAT <dbl>,
#> #   MED_MONSATFAT <dbl>, MED_ALCOHOL <dbl>