MEDI_V2 Calculation

Calculate the Mediterranean dietary index in serving size-based and 5 point scoring criteria, using given the serving sizes of foods and nutrients consumed per 1 day

Usage

MEDI_V2(
  SERV_DATA,
  RESPONDENTID,
  OLIVE_OIL_SERV_MEDI,
  FRT_SERV_MEDI,
  VEG_SERV_MEDI,
  LEGUMES_SERV_MEDI,
  NUTS_SERV_MEDI,
  FISH_SEAFOOD_SERV_MEDI,
  ALCOHOL_SERV_MEDI,
  SSB_SERV_MEDI,
  SWEETS_SERV_MEDI,
  DISCRET_FAT_SERV_MEDI,
  REDPROC_MEAT_SERV_MEDI
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

OLIVE_OIL_SERV_MEDI

The serving size of olive oil, unit= 10 g or 0.8 tbsp = 1 serving

FRT_SERV_MEDI

The serving size of All whole fruits, unit= 125 g = 1 serving (approximately 1 c)

VEG_SERV_MEDI

The serving size of All vegetables except potatoes and legumes, unit= 125 g = 1 serving (approximately 1 c of non green leafy vege; 2 cup of green leafy vege)

LEGUMES_SERV_MEDI

The serving size of legumes, including Dried beans, lentils, peas, soups (split pea), tofu, soymilk, unit= 40 g = 1 serving (about 1.5 oz)

NUTS_SERV_MEDI

The serving size of nuts and seeds, including Peanuts, almonds, sunflower seeds, cashews, walnuts, unit= 25 g = 1 serving (about 1 oz)

FISH_SEAFOOD_SERV_MEDI

The serving size of all fishes and seafoods, including Fresh-water and sea-water fish; preserved fish such as salted fish, canned fish; shellfish (squid, prawns, mollusks), unit=125 g = 1 serving (about 4 oz)

ALCOHOL_SERV_MEDI

The serving size of alcohol, including Wine, beer, "light" beer, liquor, unit= 14 g pure alcohol = 1 serving (12 ounces of beer, 5 ounces of wine, or 1.5 ounces of hard liquor, such as vodka or whiskey)

SSB_SERV_MEDI

The serving size of all dairy, including Milk, yogurt, cheese, custard, milk puddings, other milk products, unit=1 serving = 1 cup milk, 1 cup yogurt, 1 ½ ounces hard cheese (cheddar, mozzarella, Swiss, Parmesan), 1/3 cup shredded cheese, 2 ounces processed cheese (American), ½ cup ricotta cheese, 2 cups cottage cheese, 1 cup pudding made with milk, 1 cup frozen yogurt, 1 ½ cups ice cream

SWEETS_SERV_MEDI

The serving size of all sweets, including Candy, chocolate, ice cream, cookies, cakes, pies, pastries, unit= 50 g = 1 serving (about 1.5 oz ) = 1 piece of candy, 1 chocolate bar, 1 scoop ice cream, 1 cookie, 1 cake slice, 1 pie slice, 1 pastry

DISCRET_FAT_SERV_MEDI

The serving size of all discretionary fats, including Butter, margarine, mayonnaise, salad dressing, unit= 10 g = 1 serving (about 0.8 tbsp)

REDPROC_MEAT_SERV_MEDI

The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit= 150 g = 1 serving (about 5 oz)

Value

The MEDI index/score and the component scores

Examples

data("PREDIMED_trial")
MEDI_V2(SERV_DATA = PREDIMED_trial, RESPONDENTID = PREDIMED_trial$Diet_Type, OLIVE_OIL_SERV_MEDI = PREDIMED_trial$Virgin_Oliveoil, FRT_SERV_MEDI = PREDIMED_trial$Fruits, VEG_SERV_MEDI = PREDIMED_trial$Vegetables, LEGUMES_SERV_MEDI = PREDIMED_trial$Legumes, NUTS_SERV_MEDI = PREDIMED_trial$Total_nuts, FISH_SEAFOOD_SERV_MEDI = PREDIMED_trial$Fish_Seafood, ALCOHOL_SERV_MEDI = PREDIMED_trial$Alcohol, SSB_SERV_MEDI = PREDIMED_trial$Soda_Drinks, SWEETS_SERV_MEDI = PREDIMED_trial$Sweets, DISCRET_FAT_SERV_MEDI = PREDIMED_trial$Refined_Oliveoil, REDPROC_MEAT_SERV_MEDI = PREDIMED_trial$Meat)
#> # A tibble: 3 × 14
#>   RESPONDENTID MEDI_V2_ALL MEDI_V2_NOETOH MEDI_OLIVE_OIL MEDI_FRT MEDI_VEG
#>   <chr>              <dbl>          <dbl>          <dbl>    <dbl>    <dbl>
#> 1 Med_Oliveoil        42.5           35.0           5        5           5
#> 2 Med_Nuts            36.7           30.8           3.19     5           5
#> 3 Control             33.7           29.1           2.27     4.92        5
#> # ℹ 8 more variables: MEDI_LEGUMES <dbl>, MEDI_NUTS <dbl>, MEDI_FISH <dbl>,
#> #   MEDI_ALCOHOL <dbl>, MEDI_SSB <dbl>, MEDI_SWEETS <dbl>,
#> #   MEDI_DISCRET_FAT <dbl>, MEDI_REDPROC_MEAT <dbl>