PHDI Calculation

Calculate the PHDI with all gram-based variables (convert ADDED_FAT_UNSAT_SERV_PHDI, ADDED_FAT_SAT_TRANS_SERV_PHDI, ADDED_SUGAR_SERV_PHDI to percent kcal automatically using TOTALKCAL_PHDI), The Planetary Health Diet Index that quantifies adherence to the reference diet included in the EAT-Lancet report Willett et al (2019), using the serving sizes of foods and nutrients consumed per 1 day

Usage

PHDI_V2(
  SERV_DATA,
  RESPONDENTID,
  GENDER,
  TOTALKCAL_PHDI,
  WGRAIN_SERV_PHDI,
  STARCHY_VEG_SERV_PHDI,
  VEG_SERV_PHDI,
  FRT_SERV_PHDI,
  DAIRY_SERV_PHDI,
  REDPROC_MEAT_SERV_PHDI,
  POULTRY_SERV_PHDI,
  EGG_SERV_PHDI,
  FISH_SERV_PHDI,
  NUTS_SERV_PHDI,
  LEGUMES_SERV_PHDI,
  SOY_SERV_PHDI,
  ADDED_FAT_UNSAT_SERV_PHDI,
  ADDED_FAT_SAT_TRANS_SERV_PHDI,
  ADDED_SUGAR_SERV_PHDI
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients consumed per 1 day

RESPONDENTID

The unique participant ID for each participant

GENDER

The gender of the participant. 2 is female and 1 is male.

TOTALKCAL_PHDI

The total kcal

WGRAIN_SERV_PHDI

The serving size of whole grains, unit=grams/day

STARCHY_VEG_SERV_PHDI

The serving size of starchy vegetables, such as potatos, cassava, unit=grams/day

VEG_SERV_PHDI

The serving size of All vegetable except potatoes, starchy vegetables, and LEGUMES, unit=grams/day

FRT_SERV_PHDI

The serving size of All whole fruits and no fruit juice, unit=grams/day

DAIRY_SERV_PHDI

The serving size of All dairy products, unit=grams/day

REDPROC_MEAT_SERV_PHDI

The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=grams/day

POULTRY_SERV_PHDI

The serving size of Poultry, including Chicken, turkey, duck, goose, ostrich, unit=grams/day

EGG_SERV_PHDI

The serving size of Eggs with shell, unit=grams/day

FISH_SERV_PHDI

The serving size of Fish and shellfish, unit=grams/day

NUTS_SERV_PHDI

The serving size of nuts and seeds, including peanuts, tree nuts- - walnuts, almond, hazelnuts, pecan, cashew, pistachio, unit=grams/day

LEGUMES_SERV_PHDI

The serving size of Nonsoy LEGUMES, including Beans, peas, lentils, chickpeas, unit=grams/day

SOY_SERV_PHDI

The serving size of Soy products, including Tofu, tempeh, soy milk, soy yogurt, soy cheese, unit=grams/day

ADDED_FAT_UNSAT_SERV_PHDI

The serving size of Added unsaturated fat, e.g. olive soybean, rapeseed, sunflower, peanuts oil, excluding transfat, unit=grams/day

ADDED_FAT_SAT_TRANS_SERV_PHDI

The serving size of Added saturated fats and trans fat, e.g. butter, lard, coconuts oil, palm oil, unit=grams/day

ADDED_SUGAR_SERV_PHDI

The serving size of Added sugar, including the added sugar from all sweeteners and fruit juice, unit=grams/day

Value

The PHDI index/score, PHDI and its component scores