AHEI Calculation

Calculate the AHEI dietary index, Alternative Healthy Eating Index, using given the serving sizes of foods and nutrients consumed per 1 day

Usage

AHEI(
  SERV_DATA,
  RESPONDENTID,
  GENDER,
  TOTALKCAL_AHEI,
  VEG_SERV_AHEI,
  FRT_SERV_AHEI,
  WGRAIN_SERV_AHEI,
  NUTSLEG_SERV_AHEI,
  N3FAT_SERV_AHEI,
  PUFA_SERV_AHEI,
  SSB_FRTJ_SERV_AHEI,
  REDPROC_MEAT_SERV_AHEI,
  TRANS_SERV_AHEI,
  SODIUM_SERV_AHEI,
  ALCOHOL_SERV_AHEI
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

GENDER

The gender of the participant. 2 is female and 1 is male.

TOTALKCAL_AHEI

The total kcal, for adjusting sodium intake

VEG_SERV_AHEI

The serving size of All vegetable except potatoes and legume, unit=servings/day (0.5 c of vege; 1 cup of green leafy (1 cup = 236.59 g)

FRT_SERV_AHEI

The serving size of All whole fruits and no fruit juice, unit=servings/day (0.5 c of berries; 1 cup=236.59 g; 1 med fruit (1 cup = 236.59 g)

WGRAIN_SERV_AHEI

The serving size of whole grains, unit=grams/day

NUTSLEG_SERV_AHEI

The serving size of Nuts, legumes, and vegetable protein (e.g., tofu), unit=servings/day = 1 srv=1oz (28.35 g) of nuts and legume or 1 TBLSP peanut butter (15 mL), 1 cup legume = 4 oz

N3FAT_SERV_AHEI

The serving size of omega 3 fatty acid, unit=mg/day ( oz. = 28.35 g)

PUFA_SERV_AHEI

The serving size of PUFA, unit=% of energy

SSB_FRTJ_SERV_AHEI

The serving size of sugar-sweetened beverages and non-100% fruit juice, unit=servings/day = 1 ser= 8oz (1 oz. = 28.35 g)

REDPROC_MEAT_SERV_AHEI

The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=servings/day; 1 srv= 4 oz. unprocessed meat; 1.5 oz. processed meat (1 oz. = 28.35 g)

TRANS_SERV_AHEI

The serving size of trans fat, unit=% of energy

SODIUM_SERV_AHEI

The serving size of sodium, unit=mg/day per 2000 kcal

ALCOHOL_SERV_AHEI

The serving size of alcohol, including Wine, beer, "light" beer, liquor, unit=drink/day (12 oz beer; 5 oz wine; 1.5 oz spirits) 1 oz = 28.35 g

Value

The AHEI index/score, AHEI

Examples

data("AHEI_VALIDATION")
AHEI(SERV_DATA = AHEI_VALIDATION, RESPONDENTID = AHEI_VALIDATION$id, GENDER = AHEI_VALIDATION$gender, TOTALKCAL_AHEI = AHEI_VALIDATION$kcal, VEG_SERV_AHEI = AHEI_VALIDATION$vegetable, FRT_SERV_AHEI = AHEI_VALIDATION$fruit, WGRAIN_SERV_AHEI = AHEI_VALIDATION$whole_grain, NUTSLEG_SERV_AHEI = AHEI_VALIDATION$nut_legume, N3FAT_SERV_AHEI = AHEI_VALIDATION$n3_fat, PUFA_SERV_AHEI = AHEI_VALIDATION$pufa, SSB_FRTJ_SERV_AHEI = AHEI_VALIDATION$ssb_fruit_juice, REDPROC_MEAT_SERV_AHEI = AHEI_VALIDATION$red_processed_meat, TRANS_SERV_AHEI = AHEI_VALIDATION$trans_fat, SODIUM_SERV_AHEI = AHEI_VALIDATION$sodium, ALCOHOL_SERV_AHEI = AHEI_VALIDATION$alcohol)
#> # A tibble: 22 × 15
#>    RESPONDENTID GENDER AHEI_ALL AHEI_NOETOH AHEI_VEG AHEI_FRT AHEI_WGRAIN
#>           <dbl>  <dbl>    <dbl>       <dbl>    <dbl>    <dbl>       <dbl>
#>  1            1      2     42.5        40          0      0             0
#>  2            2      2     46          43.5        2      2.5           1
#>  3            3      2     55          47          4      5             2
#>  4            4      2     61.8        51.8        6      7.5           3
#>  5            5      2     66.5        56.5        8     10             4
#>  6            6      2     68.8        58.8       10     10             5
#>  7            7      2     69          59         10     10             6
#>  8            8      2     64.2        59.2       10     10             7
#>  9            9      2     59.5        59.5       10     10             8
#> 10           10      2     59.8        59.8       10     10             9
#> # ℹ 12 more rows
#> # ℹ 8 more variables: AHEI_NUTSLEG <dbl>, AHEI_N3FAT <dbl>, AHEI_PUFA <dbl>,
#> #   AHEI_SSB_FRTJ <dbl>, AHEI_REDPROC_MEAT <dbl>, AHEI_TRANS <dbl>,
#> #   AHEI_SODIUM <dbl>, AHEI_ALCOHOL <dbl>