DASHI

Calculate the DASHI dietary index (nutrient-based), Dietary Approaches to Stop Hypertension, using given the nutrients consumed per 1 day. All nutrients will be divided by (total energy/2000 kcal) to adjust energy intake.

Usage

DASHI(
  SERV_DATA,
  RESPONDENTID,
  TOTALKCAL_DASHI,
  TOTAL_FAT_DASHI,
  SAT_FAT_DASHI,
  PROTEIN_DASHI,
  CHOLESTEROL_DASHI,
  FIBER_DASHI,
  POTASSIUM_DASHI,
  MAGNESIUM_DASHI,
  CALCIUM_DASHI,
  SODIUM_DASHI
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients

RESPONDENTID

The unique participant ID for each participant

TOTALKCAL_DASHI

The total energy intake, unit = kcal/day

TOTAL_FAT_DASHI

The total fat intake, unit = g/day

SAT_FAT_DASHI

The saturated fat intake, unit = g/day

PROTEIN_DASHI

The protein intake, unit = g/day

CHOLESTEROL_DASHI

The cholesterol intake, unit = mg/day

FIBER_DASHI

The fiber intake, unit = g/day

POTASSIUM_DASHI

The potassium intake, unit = mg/day

MAGNESIUM_DASHI

The magnesium intake, unit = mg/day

CALCIUM_DASHI

The calcium intake, unit = mg/day

SODIUM_DASHI

The sodium intake, unit = mg/day

Value

The DASHI index/score

Examples

data("DASH_trial")
DASHI(SERV_DATA = DASH_trial, RESPONDENTID = DASH_trial$Diet_Type, TOTALKCAL_DASHI = DASH_trial$Kcal, TOTAL_FAT_DASHI = DASH_trial$Totalfat_Percent, SAT_FAT_DASHI = DASH_trial$Satfat_Percent, PROTEIN_DASHI = DASH_trial$Protein_Percent, CHOLESTEROL_DASHI = DASH_trial$Cholesterol, FIBER_DASHI = DASH_trial$Fiber, POTASSIUM_DASHI = DASH_trial$Potassium, MAGNESIUM_DASHI = DASH_trial$Magnesium, CALCIUM_DASHI = DASH_trial$Calcium, SODIUM_DASHI = DASH_trial$Sodium)
#> # A tibble: 5 × 12
#>   RESPONDENTID    DASHI_ALL TOTALKCAL_DASHI DASHI_TOTAL_FAT DASHI_SAT_FAT
#>   <chr>               <dbl>           <dbl>           <dbl>         <dbl>
#> 1 DASH                7.79             2100           1              0.9 
#> 2 DASH_lowSodium      8.79             2100           1              0.9 
#> 3 DASH_MedSodium      8.76             2100           1              0.9 
#> 4 DASH_HighSodium     7.79             2100           1              0.9 
#> 5 Control             0.833            2100           0.130          0.19
#> # ℹ 7 more variables: DASHI_PROTEIN <dbl>, DASHI_CHOLESTEROL <dbl>,
#> #   DASHI_FIBER <dbl>, DASHI_POTASSIUM <dbl>, DASHI_MAGNESIUM <dbl>,
#> #   DASHI_CALCIUM <dbl>, DASHI_SODIUM <dbl>