PHDI Calculation

Calculate the PHDI, The Planetary Health Diet Index that quantifies adherence to the reference diet included in the EAT-Lancet report Willett et al (2019), using the serving sizes of foods and nutrients consumed per 1 day

Usage

PHDI(
  SERV_DATA,
  RESPONDENTID,
  GENDER,
  TOTALKCAL_PHDI,
  WGRAIN_SERV_PHDI,
  STARCHY_VEG_SERV_PHDI,
  VEG_SERV_PHDI,
  FRT_SERV_PHDI,
  DAIRY_SERV_PHDI,
  REDPROC_MEAT_SERV_PHDI,
  POULTRY_SERV_PHDI,
  EGG_SERV_PHDI,
  FISH_SERV_PHDI,
  NUTS_SERV_PHDI,
  LEGUMES_SERV_PHDI,
  SOY_SERV_PHDI,
  ADDED_FAT_UNSAT_SERV_PHDI,
  ADDED_FAT_SAT_TRANS_SERV_PHDI,
  ADDED_SUGAR_SERV_PHDI
)

Arguments

SERV_DATA

The raw data file that includes all the serving sizes of foods and nutrients consumed per 1 day

RESPONDENTID

The unique participant ID for each participant

GENDER

The gender of the participant. 2 is female and 1 is male.

TOTALKCAL_PHDI

The total kcal

WGRAIN_SERV_PHDI

The serving size of whole grains, unit=grams/day

STARCHY_VEG_SERV_PHDI

The serving size of starchy vegetables, such as potatos, cassava, unit=grams/day

VEG_SERV_PHDI

The serving size of All vegetable except potatoes, starchy vegetables, and LEGUMES, unit=grams/day

FRT_SERV_PHDI

The serving size of All whole fruits and no fruit juice, unit=grams/day

DAIRY_SERV_PHDI

The serving size of All dairy products, unit=grams/day

REDPROC_MEAT_SERV_PHDI

The serving size of red and processed meats, including Beef, pork, lamb, goat, veal, sausages, bacon, salami, ham, hot dog, deli meat, unit=grams/day

POULTRY_SERV_PHDI

The serving size of Poultry, including Chicken, turkey, duck, goose, ostrich, unit=grams/day

EGG_SERV_PHDI

The serving size of Eggs with shell, unit=grams/day

FISH_SERV_PHDI

The serving size of Fish and shellfish, unit=grams/day

NUTS_SERV_PHDI

The serving size of nuts and seeds, including peanuts, tree nuts- - walnuts, almond, hazelnuts, pecan, cashew, pistachio, unit=grams/day

LEGUMES_SERV_PHDI

The serving size of Nonsoy LEGUMES, including Beans, peas, lentils, chickpeas, unit=grams/day

SOY_SERV_PHDI

The serving size of Soy products, including Tofu, tempeh, soy milk, soy yogurt, soy cheese, unit=grams/day

ADDED_FAT_UNSAT_SERV_PHDI

The serving size of Added unsaturated fat, e.g. olive soybean, rapeseed, sunflower, peanuts oil, excluding transfat, unit=percent of total kcal/day

ADDED_FAT_SAT_TRANS_SERV_PHDI

The serving size of Added saturated fats and trans fat, e.g. butter, lard, coconuts oil, palm oil, unit=percent of total kcal/day

ADDED_SUGAR_SERV_PHDI

The serving size of Added sugar, including the added sugar from all sweeteners and fruit juice, unit=percent of total kcal/day

Value

The PHDI index/score, PHDI and its component scores

Examples

data("PHDI_VALIDATION")
PHDI(SERV_DATA = PHDI_VALIDATION, PHDI_VALIDATION$id, PHDI_VALIDATION$gender, PHDI_VALIDATION$TOTALKCAL_PHDI, PHDI_VALIDATION$WGRAIN_SERV_PHDI, PHDI_VALIDATION$STARCHY_VEG_SERV_PHDI, PHDI_VALIDATION$VEG_SERV_PHDI, PHDI_VALIDATION$FRT_SERV_PHDI, PHDI_VALIDATION$DAIRY_SERV_PHDI, PHDI_VALIDATION$REDPROC_MEAT_SERV_PHDI, PHDI_VALIDATION$POULTRY_SERV_PHDI, PHDI_VALIDATION$EGG_SERV_PHDI, PHDI_VALIDATION$FISH_SERV_PHDI, PHDI_VALIDATION$NUTS_SERV_PHDI, PHDI_VALIDATION$LEGUMES_SERV_PHDI, PHDI_VALIDATION$SOY_SERV_PHDI, PHDI_VALIDATION$ADDED_FAT_UNSAT_SERV_PHDI, PHDI_VALIDATION$ADDED_FAT_SAT_TRANS_SERV_PHDI, PHDI_VALIDATION$ADDED_SUGAR_SERV_PHDI)
#> # A tibble: 26 × 19
#>    RESPONDENTID GENDER PHDI_ALL TOTALKCAL_PHDI PHDI_WGRAIN PHDI_STARCHY_VEG
#>           <dbl>  <dbl>    <dbl>          <dbl>       <dbl>            <dbl>
#>  1            1      2        0           2000           0                0
#>  2            2      2        0           2000           0                0
#>  3            3      2       14           2000           1                1
#>  4            4      2       28           2000           2                2
#>  5            5      2       42           2000           3                3
#>  6            6      2       56           2000           4                4
#>  7            7      2       70           2000           5                5
#>  8            8      2       84           2000           6                6
#>  9            9      2       98           2000           7                7
#> 10           10      2      112           2000           8                8
#> # ℹ 16 more rows
#> # ℹ 13 more variables: PHDI_VEG <dbl>, PHDI_FRT <dbl>, PHDI_DAIRY <dbl>,
#> #   PHDI_REDPROC_MEAT <dbl>, PHDI_POULTRY <dbl>, PHDI_EGG <dbl>,
#> #   PHDI_FISH <dbl>, PHDI_NUTS <dbl>, PHDI_LEGUMES <dbl>, PHDI_SOY <dbl>,
#> #   PHDI_ADDED_FAT_UNSAT <dbl>, PHDI_ADDED_FAT_SAT <dbl>,
#> #   PHDI_ADDED_SUGAR <dbl>